Move 3d geometry to separate module
