+++ APiR/CMakeLists.txt
@@ -15,6 +15,7 @@ add_subdirectory(Visual)
 add_subdirectory(BMPWriter)
 add_subdirectory(Physics)
 add_subdirectory(Geometry)
+add_subdirectory(Geometry.3D)
 add_subdirectory(Fluid)
 add_subdirectory(Window)
 add_subdirectory(Main)
