+++ APiR/Geometry/CMakeLists.txt
@@ -1,17 +1,7 @@
 set(
   SOURCES
   
-  Transformable.cpp
   Transformation.cpp
-  
-  ISurface.cpp
-  Sphere.cpp
-  Torus.cpp
-  Cylinder.cpp
-  Plane.cpp
-  Ray.cpp
-  
-  Intersection.cpp
 )
 
 set(
@@ -19,28 +9,14 @@ set(
   
   BoundingBox.h
   
-  Transformable.h
   Transformation.h
-  
-  Cylinder.h
-  ISurface.h
-  Plane.h
-  Ray.h
-  Sphere.h
-  Torus.h
-  
-  Intersection.h
 )
 
 set(
   PY
   
   BoundingBox_py.hpp
-  Intersection_py.hpp
-  ISurface_py.hpp
   geometry_py.hpp
-  Ray_py.hpp
-  Sphere_py.hpp
 )
 
 set(
