+++ APiR/Main/include/Main/SceneExamples.h
@@ -1,17 +1,5 @@
 #pragma once
-#include <Geometry/Sphere.h>
-#include <Geometry/Cylinder.h>
-#include <Geometry/Torus.h>
-
-#include <Math/Constants.h>
-
-#include <Rendering/RenderableObject.h>
 #include <Rendering/Scene.h>
-
-#include <Visual/PhongMaterial.h>
-#include <Visual/SpotLight.h>
-
-
 
 namespace ExampleScene {
   Scene OneSphere();
