+++ APiR/Rendering/include/Rendering/Camera.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <Geometry/Ray.h>
+#include <Geometry.3D/Ray.h>
 
 #include <Math/Vector.h>
